+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All significant modifications to this project will be recorded in this file. The
 
 
 
+## [1.1.0] - 2024-10-29
+
+### Added
+
+- Renumbering and reordering of a table (`reindex_table()`)
+- Solving in the form `y=mi(x-ai)` for Proposition 3.1. from [this paper](https://www.researchgate.net/publication/1893173_On_simple_arrangements_of_lines_and_pseudo-lines_in_P2_and_R2_with_the_maximum_number_of_triangles).
+- Second [gallery](https://zegalur.github.io/line-order/gallery/index_2.html)
+
 ## [1.0.0] - 2024-10-21
 
 ### Added